+++ R/summarize.R
@@ -7,6 +7,16 @@
 #'
 #' @param .data A Polars Data/LazyFrame
 #' @inheritParams mutate.RPolarsDataFrame
+#' @param .groups Grouping structure of the result. Must be one of:
+#' * `"drop_last"` (default): drop the last level of grouping;
+#' * `"drop"`: all levels of grouping are dropped;
+#' * `"keep"`: keep the same grouping structure as `.data`.
+#'
+#' For now, `"rowwise"` is not supported. Note that `dplyr` uses `.groups =
+#' NULL` by default, whose behavior depends on the number of rows by group in
+#' the output. However, returning several rows by group in `summarize()` is
+#' deprecated (one should use `reframe()` instead), which is why `.groups =
+#' NULL` is not supported by `tidypolars`.
 #'
 #' @export
 #' @examplesIf require("dplyr", quietly = TRUE) && require("tidyr", quietly = TRUE)
@@ -19,17 +29,23 @@
 #' mtcars |>
 #'   as_polars_df() |>
 #'   summarize(m_gear = mean(gear), sd_gear = sd(gear), .by = cyl)
-
-summarize.RPolarsDataFrame <- function(.data, ..., .by = NULL) {
-
+summarize.RPolarsDataFrame <- function(.data, ..., .by = NULL, .groups = "drop_last") {
   grps <- get_grps(.data, rlang::enquo(.by), env = rlang::current_env())
   mo <- attributes(.data)$maintain_grp_order
   if (is.null(mo)) mo <- FALSE
   is_grouped <- !is.null(grps)
   is_rowwise <- attributes(.data)$grp_type == "rowwise"
 
-  # do not take the groups into account, especially useful when applying across()
-  # on everything()
+  # Technically, .groups can be NULL and then the value depends on the number
+  # of rows for each group after aggregation, but returning multiple rows is
+  # deprecated so I only use those 4 values.
+  .groups <- rlang::arg_match0(.groups, values = c("drop_last", "drop", "keep", "rowwise"))
+  if (.groups == "rowwise") {
+    abort("`tidypolars` doesn't support `.groups = \"rowwise\"` for now.")
+  }
+
+  # Do not take the groups into account, especially useful when applying across()
+  # on everything().
   .data_for_translation <- select(.data, -all_of(grps))
   polars_exprs <- translate_dots(
     .data = .data_for_translation,
@@ -57,6 +73,15 @@ summarize.RPolarsDataFrame <- function(.data, ..., .by = NULL) {
   }
 
   out <- if (is_grouped && missing(.by)) {
+    grps <- switch(.groups,
+      "drop_last" = grps[-length(grps)],
+      "drop" = character(0),
+      "keep" = grps,
+      abort("Unreachable")
+    )
+    if (length(grps) == 0) {
+      return(.data)
+    }
     group_by(.data, all_of(grps), maintain_order = mo)
   } else if (isTRUE(is_rowwise)) {
     rowwise(.data)